Fix pacu region issue
This PR updates the Snapshot and LocalSnapshot classes in dsnap/snapshot.py to accept an explicit region parameter. 

When a region is provided, the code now creates a new boto3 session using that region—overriding any existing session details. This ensures that dsnap will always query AWS using the intended region, even when invoked as a library (for example, by Pacu).